Deleted on calendar and made sleep and mood on one calendar
