+++ frontend/src/components/Pages/CalendarPage.jsx
@@ -1,10 +1,8 @@
 import React, { useState, useEffect } from 'react';
-import Calendar from 'react-calendar';
-import 'react-calendar/dist/Calendar.css';
 import { Vortex } from '../ui/vortex';
 import { Navbar2 } from '../Navbar2';
-import MoodCalendar from '../MoodCalendar';
-import SleepCalendar from '../SleepCalendar';
+import Calendar from '../Calendar';
+
 const CalendarPage = () => {
 
     return (
@@ -23,7 +21,7 @@ const CalendarPage = () => {
           <div
             style={{
               width: "100%",
-              maxWidth: "60%",
+              maxWidth: "500px",
               padding: "40px",
               backgroundColor: "black",
               borderRadius: "12px",
@@ -32,9 +30,7 @@ const CalendarPage = () => {
                
             }}
           >
-           
-           <MoodCalendar/>
-           <SleepCalendar/>
+           <Calendar/>
         </div>
         </div>
         </Vortex>
